+++ mapwize/src/main/java/io/mapwize/mapwize/MWZPlaceList.java
@@ -1,10 +1,14 @@
 package io.mapwize.mapwize;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonSetter;
 
 import java.util.List;
 import java.util.Map;
 
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class MWZPlaceList implements MWZDirectionPoint, MWZSearchable {
 
     private String identifier;